refactor: move table helpers to shared module
