+++ messung/templates/messung/projekte_page.html
@@ -220,5 +220,5 @@
     });
   }
 </script>
-<script src="{% static 'js/projekte.js' %}"></script>
+<script type="module" src="{% static 'js/projekte.js' %}"></script>
 {% endblock %}
